editPage 하단 여백추가 / navigation에 loginData가 비어있으면 loginPage로 보내는 단순 조건문 추가 / service에 작성된 API호출 함수에 예외처리 추가, 디테일한 추가는 errCode에 따라 하고싶지만 백엔드쪽 작업처리가 안되있으므로 천천히 추가 할 예정

+++ app/src/service/deleteOneEdit.ts
@@ -6,6 +6,12 @@ export const deleteOneEdit = async (body: {
   dayID: string;
 }) => {
   try {
+    if (!body.email || !body.monthID || !body.dayID) {
+      alert(
+        "삭제에 필요한 데이터가 누락되었습니다, 같은 에러가 반복된다면 관리자에게 문의하세요"
+      );
+      return;
+    }
     const response = await api.delete("/edit/deleteOneEdit", { data: body });
 
     return response;
